rename from clust_subsetter to clust_formatter
